Switch from Qt 4 to Qt 5
I've only tested on this on Debian Jessie with Qt 5.3. Panning and
rotating the model viewports cause the application to lag. Though
posts online suggest it is caused by a Qt bug where mouse move events
are not combined correctly on X11 and the application's event handler
calls updateGL() each time. It was reportedly fixed in Qt 5.6.

https://bugreports.qt.io/browse/QTBUG-40889
https://bugs.archlinux.org/task/47650
https://bugs.scribus.net/view.php?id=13363