Fix crash when opening Paint Texture
Opening Menu -> Materials -> Paint Texture... would sometimes cause
OpenGL library to crash when TextureWidget::setTexture() was called
due to wrong (or no?) OpenGL context setup.

TextureWidget::initializeGL() wasn't called beforehand either so the
OpenGL texture handle was 0.

This may not of crashed in all situations but it was reproduceable
(at least on a single machine).