Update Qt ui files, remove embedded zoom icons
I opened and saved all the Qt ui files in Qt Designer 5.3. I had to fix
ExtendSelection class name in animwin.ui and iqeprompt.ui

The zoom icons were removed that use to be embedded in some of the ui
files. It was depreciated in Qt 4 and doesn't work in Qt 5.10 any more
(and possibly earlier versions, worked in 5.3).

There was also a dithered src/pixmaps/texframe.xpm embedded with all the
TextureFrame widgets but it seems to only been used for display in
Qt Designer.