Rename to Maverick Model 3D
This is the title I used for my Misfit Model 3D fork in 2011. I'm not
sure why I didn't rename this fork started in 2014 that included the
full code history and kept automake/configure.

The alternative is to display Misfit Model 3D (clover.moe fork) in the
title bar, Windows installer, and everywhere else. That would probably
be annoying for people to type out. People would probably just refer to
the original and my fork by version numbers or something which would be
a problem if Misfit Model 3D ever resumes development.