Support loading/saving Milkshape 3D 1.8.5 BETA 1 MS3D models
Vertex bone joint weights were incorrectly parsed due to MM3D clamping
subversion 3 to 2. There is a new 32 bit field added in ms3d vertex
subVersion 3. It's not actually document what the value is suppose to be.

The unknown field is currently treated the same as suvVersion 2's
"extra" field. This may not be a good idea.