Fix new transparent texture set for material not being drawn
In View -> 3D Alpha Blend mode, there is a static bsp tree with all
transparent triangles. Regular model::draw() skips transparent textures.

Changing a material to use a new transparent texture did not update the
BSP tree so neither BSP tree or regular model::draw() rendered the group
with the new transparent texture.

This did not affect models that had a transparent texture when they were
loaded or changing group to use a material with a transparent texture.