Fix gimbal lock detection when converting matrix to Euler angles
Catch gimbal lock (need different method for getting yaw and roll
angles) when pitch is +90 or -90 degrees. Before it was only caught
if sinPitch was greater than 1.0 or less than -1.0 due to double
precision error.

This affected the exported SMD bone joints if pitch was +90 or -90
degrees. Hit or miss gimbal lock detection cause inconsistent rotation
for joints keyframes; even when all the frames were suppose to be
the same rotation.