Make Convert To Frame window convert multiple animations at once
The Convert To Frame window allows converting skeletal animations to
frame animations. It would prompt for each animation which was annoying
when converting 25 skeletal animations for a Quake 3 player to frame
animations for exporting as MD3.

I changed the Convert To Frame window to displayed all the selected
skeletal animations in a table with edit fields for frame animation
name and frame count. This allows all frame animations to be edited
at once and converted with a single click.