Switch to Qt 5.4 QOpenGLWidget
Switch from QGLWidget to QOpenGLWidget introduced in Qt 5.4.
QOpenGLWidget requires frame buffer objects so this change may cause
MM3D to no longer work on some systems. It's the modern supported
method and will make it easier to share textures and vertex buffers
between all models viewports in a window in the future.

Refactor TextureWidget::setTexture() to fix an issue with texture
coord editor being white first time it's opened. setTexture() was
called before initializeGL() causing texture to be created on wrong
OpenGL context.

Grid size in modelviewport used QGLWidget::renderText but it had to
be manually enabled in config file because it's broken in some cases.
Now it's completely disabled since renderText is not supported by
QOpenGLWidget.