Fix cross product in setRotationToPoint
Quaternion::setRotationToPoint is used for bone joint rendering. It has
an in-line copy of vector 3 cross project. It accessed the wrong index
in a Vector class but it doesn't seem to make a difference. Rotation Y
axis ("B" variable) seems to always be set to zero with or without this
change.