objfilter: Fix crash for out of bounds texcoord index
If a OBJ model specifies texcoord indexes for faces but does not provide
any texcoords ("vt 0.0 0.0" lines) it would crash. Ignore texcoords for
the face instead.

Invalid OBJ texcoord index resulted in out of bounds access of m_uvList
vector which results in a SEGFAULT.

Also add an out of bounds check for OBJ vertex index. Though this did not
result in a crash as Model class range checks it. It may of resulted in
an odd error from addTriangleToGroup() though.

Reported by ZaRR.