objfilter: Fix writeStripped() flow for OBJ writing
The unlikely case of writing a number and decimal followed by a non-digit
would proceed parsing incorrecly. Writing "1. 1.000000" does not remove
trailing zeros but still writes all text.

The only time this could occur is when writting texture filename for map_Kd
which should not be removing trailing zeros in the first place.

Reported Mick Pearson.