md3filter: Add export dialog
Saving as a player model is now an export dialog option instead of a pop-up
yes, no, cancel window.

Saving animation.cfg is a separate option from player model. It can be used
for player and non-player models. (The default checkbox state in the export
dialog matches saving player model.)