mm3dfilter: Fix using getVertexCount() for points
When loading .mm3d model the 'weighted influence' section is suppose
to replace the previously loaded primary joint influences.

Vertex count was used for removing influences from the points so if
there was more points than vertexes then some legacy primary joint
influences would remain.

However there should be a weighted influence for the primary joint
which will then override the influence that was supposed to be removed.

I don't expect this change to make a difference.