Fix Animation Sets -> Split to be start frame num
The Split Animation window takes a frame number. It was documented to
be the first frame of the "second" animation but due to off-by-one
error the specified frame number was the last frame of the "first"
animation.

The GUI uses 1-based frame numbers but the internal API uses 0-based
frame numbers.