Fix copycmd arguments for writeFile()
Enum value WO_ModelNoPrompt (2) was passed to bool exportModel. This
wasn't an issue.

exportModel=true enabled filters that may not support loading models.
This isn't an issues as the copied content was always written as .mm3d
which is not export only. Using default option WO_ModelDefault is the
same as WO_ModelNoPrompt, so no change there either.