Defer changing frame count in animation mode
In animation mode with an animation with 50 frames and user types 100
the animation was truncated to 1 frame then filled with duplicate frames
to fill 10 and then 100.

In animation mode if the frame count is edited wait until user presses
enter or clicks outside of the frame count spin box to apply the change.