Fix listing files on Windows with ".." directory in path
Visual Studio may include ".." directory in GetModuleFileNameW()
for the executable name. It's used for shared plugins directory
and was causing FindFirstFileW() to fail with error 0x7b
ERROR_INVALID_NAME.

Other functions may not support ".." so just resolve ".." in all
paths that are passed to the Win32 API.

Reported by Mick Pearson.