Fix getRelativePath() with different Windows drive letters
Setting a texture for a material on a different Windows drive than
the model worked but saving and reopening a .mm3d model would not
be able to load the texture due to an invalid relative path.

With base "A:\Models\" and path "B:\Images\File.png" the relative
path was returned as "./../../:/Images/File.png". Instead return
the absolute path with the drive letter ("B:/Images/File.png").