Enable scaling window coords for HiDPI
Enable flag to scale window coords from pixels to points for HiDPI.
This fixes a bunch of stuff being too small, like the status bar height,
toolbar icons, and text boxes.

Strangely resizeGL() always receives points (unrelated to
AA_EnableHighDpiScaling) and needs to be scaled. I swear this was pixels
before when I fixed the rest of the code for HiDPI. So I'm not sure if
Qt changed behavior or my memory is wrong. If Qt changed behavior, this
should have a version check.