Disable text editing for frame count in animation mode
Don't allow text editing in the frame count QSpinBox. It updates the
frame count immediately. 9 -> 10 will truncate to 1 frame and then add
empty frames.

Deferring it until losing focus is confusing. I don't really want to
add an apply button as it's feels like an unnecessary step when using
the +/- buttons. So just disable text editing.

It can be set with text editing in the New Animation and Edit Animation
windows.